Install to build directory instead of system
Install the binaries along with their SPV files to the cmake build
directory instead of a random system install directory. This mirrors how
the compute-samples are installed. This is useful because the working
directory when executing the binaries has to have the SPV files in it so
the binaries can find them.

Signed-off-by: Lisanna Dettwyler <lisanna.dettwyler@intel.com>