Pre-merge test execution
Tests are executed concurrently against fulsim-gen9 in containers run
on CaaS. Tests are executed on a per-binary basis with a progress
timeout of 60 seconds (default, configurable in test_config.yml). Merge
requests are not gated by test passing - rather, it's up to individual
developers to interpret the test execution results and make informed
decisions about the correctness of their patch.

Closes VLCLJ-406
Closes VLCLJ-402
Closes VLCLJ-403
Closes VLCLJ-456