Granular test execution with regression reports
Pre-merge execution was significantly rewritten to support granular
concurrent execution of tests at the testcase level. The code was
largely rewritten in python, re-implementing the concurrent execution,
container launch, timeout, and result reporting.

JUnit reports are generated that can be consumed by GitLab for the
purposes of result presentation and regression analysis.

Closes VLCLJ-404
Closes VLCLJ-482
Closes VLCLJ-485