CI refactor supporting use of pylzt module
Use the external pylzt module to provide the regression execution and
AUB generation functionality. Moving these functions to an external
module will allow it to be used by other projects.

The runtime-ubuntu1804 image is built and published for commits to
master, with the commit SHA as the tag, allowing it to also be easily
referenced by other projects.

Some of the docker image caching and Dockerfile organization was
reworked.

Related to VLCLJ-597